% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCSVtoP2distance.R
\name{loadCSVtoP2distance}
\alias{loadCSVtoP2distance}
\title{Load a CSV file into a matrix object}
\usage{
loadCSVtoP2distance(
  path,
  header = TRUE,
  sep = "\\t",
  dec = ".",
  quote = "\\"",
  na.strings = "NA",
  fileEncoding = "",
  encoding = "unknown"
)
}
\arguments{
\item{path}{The path of the file from which the data are to be read. Each
row of the table appears as one line of the file.}

\item{header}{A logical value indicating whether the file contains the
names of the variables as its first line.}

\item{sep}{The field separator character.}

\item{dec}{The character used in the file for decimal points.}

\item{quote}{The set of quoting characters. To disable quoting
altogether, use \code{quote = ""}.}

\item{na.strings}{A character vector of strings to be interpreted as \code{NA}
values.}

\item{fileEncoding}{Character string: if non-empty, declares the encoding
used in the file so the character data can be re-encoded. See the
"Encoding" section of \code{?file}.}

\item{encoding}{Encoding to be assumed for input strings.}
}
\value{
A matrix containing the data from the CSV file, ready to be
passed to \code{\link[=p2distance]{p2distance()}}.
}
\description{
Reads a text file and converts it into the matrix object expected by
\code{\link[=p2distance]{p2distance()}}.
}
\details{
The first column of the file is used to set the row names of the
resulting matrix (typically the name of each spatial entity), and is then
removed from the data itself. Internally this uses \code{\link[utils:read.table]{utils::read.table()}}
to read the file.
}
\seealso{
\code{\link[=p2distance]{p2distance()}}
}
