% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReferenceVector.R
\name{makeReferenceVector}
\alias{makeReferenceVector}
\title{Make a reference vector}
\usage{
makeReferenceVector(X, reference_vector_function = min)
}
\arguments{
\item{X}{A data frame or matrix with \code{n} rows (entities) and \code{m} columns
(variables).}

\item{reference_vector_function}{Function used to build the reference
vector. Common choices are \code{min}, \code{max}, \code{mean}, \code{median}, etc. See
\code{\link[=apply]{apply()}} for details. \code{min} is the default.}
}
\value{
A vector with the reference value (base reference) for each
variable. Its length equals the number of variables.
}
\description{
A reference vector must be defined for each partial indicator so as to
compare different spatial entities. This vector is used by
\code{\link[=p2distance]{p2distance()}} (as the base reference) to calculate the distance of each
spatial entity from this reference base. It is common to use the minimum
value as the base reference (see references).
}
\examples{
## Create a data frame of 3 variables (indicators) for 6 entities (rows)
dat <- data.frame(
  x1 = c(10, 12, 13, 14, 12, 11),
  x2 = c(40, 51, 61, 68, 34, 44),
  x3 = c(0.43, 0.41, 0.39, 0.55, 0.60, 0.38)
)

## Calculate the reference vector using the minimum value of each variable
makeReferenceVector(dat, reference_vector_function = min)

}
\references{
Peña, J. B. (1977). \emph{Problemas de la medición del bienestar y conceptos
afines (una aplicación al caso Español)}. Madrid: INE.

Peña, J. B. (2009). La medición del bienestar social: una revisión
crítica. \emph{Estudios de Economía Aplicada}, 27(2), 299-324.

Somarriba, N. (2008). \emph{Aproximación a la medición de la calidad de vida
en la Unión Europea}. Doctoral Thesis. University of Valladolid.

Zarzosa, P. (1992). \emph{Aproximación a la medición del bienestar social,
estudio de la idoneidad del indicador sintético Distancia P2}. Doctoral
Thesis. University of Valladolid.
}
\seealso{
\code{\link[=p2distance]{p2distance()}}
}
