% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2distance-package.R
\docType{package}
\name{p2distance-package}
\alias{p2distance-package}
\title{p2distance: Welfare's Synthetic Indicator}
\description{
The welfare's synthetic indicator (see references) provides an ideal
tool for measuring multi-dimensional concepts such as welfare,
development, living standards, etc. It enables information from various
indicators to be aggregated into a single synthetic measure. This
synthetic indicator, also called \eqn{P_2} distance, overcomes several
limitations of other traditional approaches, such as Principal
Component Analysis (PCA) or Data Envelopment Analysis (DEA), and offers
good statistical properties, allowing both inter-spatial and
inter-temporal comparisons.
}
\details{
The main function of the package is \code{\link[=p2distance]{p2distance()}}.
}
\references{
Peña, J. B. (1977). \emph{Problemas de la medición del bienestar y conceptos
afines (una aplicación al caso Español)}. Madrid: INE.

Peña, J. B. (2009). La medición del bienestar social: una revisión
crítica. \emph{Estudios de Economía Aplicada}, 27(2), 299-324.

Somarriba, N. (2008). \emph{Aproximación a la medición de la calidad de vida
en la Unión Europea}. Doctoral Thesis. University of Valladolid.

Zarzosa, P. (1992). \emph{Aproximación a la medición del bienestar social,
estudio de la idoneidad del indicador sintético Distancia P2}. Doctoral
Thesis. University of Valladolid.

Zarzosa, P. (1996). \emph{Aproximación a la medición del Bienestar social}.
Valladolid: Universidad de Valladolid.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajpelu/p2distance}
  \item \url{https://ajpelu.github.io/p2distance/}
  \item Report bugs at \url{https://github.com/ajpelu/p2distance/issues}
}

}
\author{
\strong{Maintainer}: Antonio J. Pérez-Luque \email{ajpelu@gmail.com} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID})

Authors:
\itemize{
  \item Antonio J. Pérez-Luque \email{ajpelu@gmail.com} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID})
  \item Ramón Pérez-Pérez
}

Other contributors:
\itemize{
  \item Ricardo Moreno [contributor]
  \item Francisco J. Bonet-García [contributor]
}

}
\keyword{internal}
