\encoding{latin1}
\name{p2distance-package}
\alias{p2distance-package}
\docType{package}
\title{
Welfare's Synthetic Indicator
~~ p2distance ~~
}
\description{
The welfare's synthetic indicator (see references) provides an ideal tool for measuring multi-dimensional concepts such as welfare, development, living standards, etc. It enables information from the various indicators to be aggregated into a single synthetic measure. This synthetic indicator, also called \eqn{P_{2}} distance, overcomes several limitations of others traditional approach as Principal Components Analysis (PCA), Data Envelopment Analysis (DEA), etc. and offers some good statistical propierties and allows inter-spatial an inter-temporal comparisons.  
}
\details{
\tabular{ll}{
Package: \tab p2distance\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2012-05-02\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
~~ p2distance is the main function ~~
}
\author{
A.J. Perez-Luque; R. Moreno; R. Perez-Perez and F.J. Bonet
}
\references{
Pena, J. B. (1977). \emph{Problemas de la medición del bienestar y conceptos afines (una aplicación al caso Español)}. Madrid: INE.

Pena, J. B. (2009). La medición del bienestar social: una revisión crítica. \emph{Estudios de Economía Aplicada}, \bold{27(2)}, 299--324.

Somarriba, N. (2008). \emph{Aproximación a la medición de la calidad de vida en la Unión Europea}. Doctoral Thesis. University of Valladolid. \url{http://www.eumed.net/tesis/2010/mnsa/index.htm}

Zarzosa, P. (1992). \emph{Aproximación a la medición del bienestar social, estudio de la idoneidad del indicador sintético Distancia P2}. Doctoral Thesis.  University of Valladolid.

Zarzosa, P. (1996). \emph{Aproximación a la medición del Bienestar social}. Valladolid: University of Valladolid. 248 pp. 
}

\keyword{p2distance}
\keyword{synthetic indicators}
